Add fixes to pools
